+++ ovos_utils/system.py
@@ -3,6 +3,25 @@ import sys
 import os
 import subprocess
 import re
+from enum import Enum
+import platform
+from ovos_utils.display import can_display, is_gui_installed
+
+
+class MycroftRootLocations(str, Enum):
+    PICROFT = "/home/pi/mycroft-core"
+    BIGSCREEN = "/home/mycroft/mycroft-core"
+    OVOS = "/home/mycroft/mycroft-core"  # TODO ovos here
+    OLD_MARK1 = "/opt/venvs/mycroft-core/lib/python3.4/site-packages/"
+    MARK1 = "/opt/venvs/mycroft-core/lib/python3.7/site-packages/"
+    MARK2 = "/home/mycroft/mycroft-core"  # TODO mark2 here
+
+
+def search_mycroft_core_location():
+    for p in MycroftRootLocations:
+        if os.path.isdir(p):
+            return p
+    return None
 
 
 def get_desktop_environment():
@@ -18,13 +37,16 @@ def get_desktop_environment():
         desktop_session = os.environ.get("DESKTOP_SESSION")
         if desktop_session is not None:  # easier to match if we doesn't have  to deal with character cases
             desktop_session = desktop_session.lower()
-            if desktop_session in ["gnome", "unity", "cinnamon", "mate", "xfce4", "lxde", "fluxbox",
-                                   "blackbox", "openbox", "icewm", "jwm", "afterstep", "trinity", "kde"]:
+            if desktop_session in ["gnome", "unity", "cinnamon", "mate",
+                                   "xfce4", "lxde", "fluxbox",
+                                   "blackbox", "openbox", "icewm", "jwm",
+                                   "afterstep", "trinity", "kde"]:
                 return desktop_session
             # Special cases
             # Canonical sets $DESKTOP_SESSION to Lubuntu rather than LXDE if using LXDE.
             # There is no guarantee that they will not do the same with the other desktop environments.
-            elif "xfce" in desktop_session or desktop_session.startswith("xubuntu"):
+            elif "xfce" in desktop_session or desktop_session.startswith(
+                    "xubuntu"):
                 return "xfce4"
             elif desktop_session.startswith("ubuntu"):
                 return "unity"
@@ -63,9 +85,42 @@ def is_process_running(process):
     except:  # Windows
         s = subprocess.Popen(["tasklist", "/v"], stdout=subprocess.PIPE)
     for x in s.stdout:
-        if re.search(process, x):
+        if re.search(process, x.decode("utf-8")):
             return True
     return False
+
+
+def get_platform_fingerprint():
+    # TODO hostname
+
+    # TODO avoid circular import
+    from ovos_utils.configuration import read_mycroft_config
+    conf = read_mycroft_config()
+    listener_conf = conf.get("listener", {})
+    skills_conf = conf.get("skills", {})
+    return {
+        "platform": platform.platform(),
+        "enclosure": conf.get("enclosure", {}).get("platform"),
+        "python_version": platform.python_version(),
+        "system": platform.system(),
+        "version": platform.version(),
+        "arch": platform.machine(),
+        "data_dir": conf.get("data_dir"),
+        "msm_skills_dir": skills_conf.get("msm", {}).get("directory"),
+        "release": platform.release(),
+        "node": platform.node(),
+        "desktop_env": get_desktop_environment(),
+        "ipc_path": conf.get("ipc_path"),
+        "input_device_name": listener_conf.get("device_name"),
+        "input_device_index": listener_conf.get("device_index"),
+        "default_audio_backend": conf.get("Audio", {}).get("default-backend"),
+        "priority_skills": skills_conf.get("priority_skills"),
+        "backend_url": conf.get("server", {}).get("url"),
+        "mycroft_core_location": search_mycroft_core_location(),
+        "can_display": can_display(),
+        "is_gui_installed": is_gui_installed(),
+        "pulseaudio_running": is_process_running("pulseaudio")
+    }
 
 
 def ntp_sync():
@@ -95,4 +150,3 @@ def ssh_disable():
     # Permanently block SSH access from the outside
     subprocess.call('sudo systemctl stop ssh.service', shell=True)
     subprocess.call('sudo systemctl disable ssh.service', shell=True)
-
